+++ typescript/packages/subsurface-viewer/src/layers/axes/boxLayer.ts
@@ -9,7 +9,7 @@ import type {
     ExtendedLayerProps,
 } from "../utils/layerTools";
 
-import { testPrecision } from "../shader_modules/test-precision/test-precision";
+import { precisionForTests } from "../shader_modules/test-precision/precisionForTests";
 
 import fragmentShader from "./box.fs.glsl";
 import vertexShader from "./box.vs.glsl";
@@ -47,7 +47,7 @@ export default class BoxLayer extends Layer<BoxLayerProps> {
             ...super.getShaders({
                 vs: vertexShader,
                 fs: fragmentShader,
-                modules: [project32, testPrecision],
+                modules: [project32, precisionForTests],
             }),
             uniforms: { uColor: Array.from(color) },
             geometry: new Geometry({
