+++ typescript/packages/subsurface-viewer/src/layers/northarrow/northArrow3DLayer.ts
@@ -10,7 +10,7 @@ import type { Device } from "@luma.gl/core";
 import { Vector3 } from "@math.gl/core";
 import type { ExtendedLayerProps } from "../utils/layerTools";
 
-import { testPrecision } from "../shader_modules/test-precision/test-precision";
+import { precisionForTests } from "../shader_modules/test-precision/precisionForTests";
 
 import fragmentShader from "./northarrow.fs.glsl";
 import vertexShader from "./northarrow.vs.glsl";
@@ -103,7 +103,7 @@ export default class NorthArrow3DLayer extends Layer<NorthArrow3DLayerProps> {
             ...super.getShaders({
                 vs: vertexShader,
                 fs: fragmentShader,
-                modules: [project32, testPrecision],
+                modules: [project32, precisionForTests],
             }),
             uniforms: { uColor: Array.from(color) },
             geometry: new Geometry({
