+++ typescript/packages/subsurface-viewer/src/layers/piechart/pieChartLayer.ts
@@ -14,7 +14,7 @@ import type {
 } from "../utils/layerTools";
 import { createPropertyData } from "../utils/layerTools";
 
-import { testPrecision } from "../shader_modules/test-precision/test-precision";
+import { precisionForTests } from "../shader_modules/test-precision/precisionForTests";
 
 import fragmentShader from "./piechart.fs.glsl";
 import vertexShader from "./piechart.vs.glsl";
@@ -158,7 +158,7 @@ export default class PieChartLayer extends Layer<PieChartLayerProps<PiesData>> {
             ...super.getShaders({
                 vs: vertexShader,
                 fs: fragmentShader,
-                modules: [project32, picking, testPrecision],
+                modules: [project32, picking, precisionForTests],
             }),
             geometry: new Geometry({
                 topology: "triangle-list",
