+++ typescript/packages/subsurface-viewer/src/layers/points/privatePointsLayer.ts
@@ -6,7 +6,7 @@ import { GL } from "@luma.gl/constants";
 
 import type { ShaderModule } from "@luma.gl/shadertools";
 
-import { testPrecision } from "../shader_modules/test-precision/test-precision";
+import { precisionForTests } from "../shader_modules/test-precision/precisionForTests";
 
 import vs from "./point.vs.glsl";
 
@@ -28,7 +28,11 @@ export class PrivatePointsLayer extends ScatterplotLayer<
         // use object.assign to make sure we don't overwrite existing fields like `vs`, `modules`...
         return Object.assign({}, superShaders, {
             vs: vs,
-            modules: [...superShaders.modules, pointsUniforms, testPrecision],
+            modules: [
+                ...superShaders.modules,
+                pointsUniforms,
+                precisionForTests,
+            ],
         });
     }
