+++ typescript/packages/subsurface-viewer/src/layers/shader_modules/test-precision/precisionForTests.ts
@@ -1,10 +1,10 @@
 import type { ShaderModule } from "@luma.gl/shadertools";
 
-import { PRECISION } from "./precision-shaders-glsl";
+import { PRECISION } from "./precisionForTests-shaders-glsl";
 
 /** In Phong shading, the normal vector is linearly interpolated across the surface of the polygon from the polygon's vertex normals. */
-export const testPrecision: ShaderModule = {
-    name: "testPrecision",
+export const precisionForTests: ShaderModule = {
+    name: "precisionForTests",
     vs: PRECISION,
     fs: PRECISION,
 };
